@@ -19,19 +19,19 @@
         }
     },
     "require": {
-        "php": ">=8.0",
-        "hyperf/config": "~3.0",
-        "hyperf/db-connection": "~3.0",
-        "hyperf/di": "~3.0",
-        "hyperf/framework": "~3.0",
-        "hyperf/validation": "~3.0"
+        "php": ">=8.1",
+        "hyperf/config": "~3.1",
+        "hyperf/db-connection": "~3.1",
+        "hyperf/di": "~3.1",
+        "hyperf/framework": "~3.1",
+        "hyperf/validation": "~3.1"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
         "hyperf/testing": "~3.0",
         "mockery/mockery": "^1.0",
         "phpstan/phpstan": "^1.0",
-        "swoole/ide-helper": "^4.5"
+        "swoole/ide-helper": "^5.0"
     },
     "suggest": {
         "swow/swow": "Required to create swow components."
